pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cykit"]

[project]
name = "cykit"
version = "4.0.0"
description = "Emotiv EEG headset data acquisition (cross-platform)"
requires-python = ">=3.10"
dependencies = [
    "pycryptodomex>=3.19.0",
    "bleak>=0.21.0",
    "pyusb>=1.2.0",
]

[project.scripts]
cykit = "cykit.main:cli"
